deduplicated parse_calldata and parse_code
